Fixes for the 64-bit architecture I've been really messed up these days... The trauma that I got over food and the work I had to get done in this engine, added up with school's pression made be just want to quit it... My mental and physical state are spiking down again, it can't be... Now I'm thinking of going to school this afternoon I don't know.. d3d11 will be worked on very soon too

+++ src/core/drivers/sdl_renderer/sdlrenderer_draw.c
@@ -144,7 +144,7 @@ void sresdlrenderer_flush_queueinstances2(void* _inst, const sre_RenderInstance2
                 NULL,
                 &vertices->x, sizeof(sre_vec2f),
                 (const SDL_Color*)&instance->color, 0,
-                NULL, 0, point_count, NULL, 0, 0
+                NULL, 0, (int)point_count, NULL, 0, 0
             );
         } break;
         case SRE_DRAW2_STRIP:
@@ -168,8 +168,8 @@ void sresdlrenderer_flush_queueinstances2(void* _inst, const sre_RenderInstance2
                 NULL,
                 &vertices->x, sizeof(sre_vec2f),
                 (const SDL_Color*)&instance->color, 0,
-                NULL, 0, point_count,
-                indices, indice_count, 2
+                NULL, 0, (int)point_count,
+                indices, (int)indice_count, 2
             );
             SDL_stack_free(indices);
         } break;
